+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Next
+## [v2.4.1](https://github.com/vazco/uniforms/tree/v2.4.1) (2019-10-02)
 
+- **Fixed:** Name conflict of `type` in `JSONSchemaBridge`. [\#604](https://github.com/vazco/uniforms/issues/604)
 - **Fixed:** Parsing of `dataPath` in `JSONSchemaBridge`. [\#593](https://github.com/vazco/uniforms/issues/593)
 
 ## [v2.4.0](https://github.com/vazco/uniforms/tree/v2.4.0) (2019-08-28)